hash-types: re-structure modules to hold all relevant data structures
