+++ compiler/hash-typecheck/src/exhaustiveness/fields.rs
@@ -6,7 +6,10 @@
 //! from a provided [PatCtx]. [FieldOps] defines methods that operate on
 //! [Fields] with the typechecker context available for reading and creating
 //! [DeconstructedPat](super::deconstruct::DeconstructedPat)s.
-use hash_types::{terms::TermId, Level1Term, NominalDef, StructFields, Term, TupleTy};
+use hash_types::{
+    nominals::{NominalDef, StructFields},
+    terms::{Level1Term, Term, TermId, TupleTy},
+};
 use hash_utils::store::Store;
 use itertools::Itertools;
