+++ compiler/hash-typecheck/src/traverse/scopes.rs
@@ -3,7 +3,7 @@
 
 use hash_ast::{ast, visitor::AstVisitor};
 use hash_source::identifier::Identifier;
-use hash_types::{scope::ScopeId, ScopeKind};
+use hash_types::scope::{ScopeId, ScopeKind};
 
 use super::visitor::TcVisitor;
 use crate::{diagnostics::error::TcResult, ops::AccessToOps};
